Only use prebuilt packages if a kerncache was found
This patch modifies the genkernel callback to only include `-kb`
to build and use prebuilt packages for extra kernel modules, if
the a prebuilt kernel was found.

Importantly, this patch will build packages that depend on the
version of the built kernel to be rebuilt if the version of the
kernel used by a catalyst run is upgraded.

This can happen if `boot/kernel/gentoo/packages` binary packages
depend on the specific kernel version.

e.g. if they are an out-of-tree kernel module, like sys-fs/zfs-kmod